+++ pkg/providers/dns/testdata/TestResDnsZone/create_multisigner_without_enabled.tf
@@ -6,12 +6,11 @@ resource "akamai_dns_zone" "multi_signer_test_zone" {
   contract       = "ctr1"
   zone           = "multisignerexampleterraform.io"
   type           = "primary"
-  comment        = "This is an updated test zone with multi-signer DNSSEC"
+  comment        = "This is a test zone with multi-signer DNSSEC"
   sign_and_serve = true
   group          = "grp1"
 
   multi_provider_dnssec {
-    enabled = true
     webhook = "https://example.com/webhook"
   }
 }